+++ NormalMapGeneratorTool/src/FrameBufferSystem.h
@@ -10,7 +10,7 @@ private:
 public:
 	FrameBufferSystem();
 	//Initialize frame buffer with certain resolution
-	void init(const glm::vec2& windowRes, const glm::vec2& maxBufferResolution);
+	void init(const glm::ivec2& windowRes, const glm::ivec2& maxBufferResolution);
 	//Initialize frame buffer with certain resolution
 	void init(int windowWidth, int windowHeight, int maxBufferWidth, int maxBufferHeight);
 	//Make the this the new bound frame buffer
@@ -26,15 +26,15 @@ public:
 	//Change resolution of existing frame buffer
 	void updateTextureDimensions(int windowWidth, int windowHeight) noexcept;
 	//Change resolution of existing frame buffer
-	void updateTextureDimensions(const glm::vec2 &windowRes) noexcept;
+	void updateTextureDimensions(const glm::ivec2 &windowRes) noexcept;
 	//Get the frame buffer id
 	unsigned int getFrameBufferId() const;
 	//Get the FBO that is current bound
 	static unsigned int getCurrentlyBoundFBO() noexcept;
 	static void blit(const FrameBufferSystem& source, const FrameBufferSystem& destination,
-		const glm::vec2& srcStartCoord, const glm::vec2& srcEndCoord, const glm::vec2& destStartCoord, 
-		const glm::vec2& destEndCoord)noexcept;
-	static void blit(const FrameBufferSystem& source, const FrameBufferSystem& destination, const glm::vec2 screenRes)noexcept;
+		const glm::ivec2& srcStartCoord, const glm::ivec2& srcEndCoord, const glm::ivec2& destStartCoord,
+		const glm::ivec2& destEndCoord)noexcept;
+	static void blit(const FrameBufferSystem& source, const FrameBufferSystem& destination, const glm::ivec2 screenRes)noexcept;
 	static void bindDefaultFrameBuffer()noexcept;
 	~FrameBufferSystem();
 };
