+++ NormalMapGeneratorTool/src/TextureLoader.h
@@ -13,12 +13,12 @@ public:
 	TextureManager(const TextureManager&) = delete;
 	static void getRawImageDataFromFile(const std::string& path, std::vector<unsigned char>& data, int &width, int &height, bool flipImage);
 	static void getTextureDataFromFile(const std::string & path, TextureData & textureData);
-	static glm::vec2 getImageDimensions(const std::string& path);
+	static glm::ivec2 getImageDimensions(const std::string& path);
 	static unsigned int createTextureFromFile(const std::string& path, bool linearColourSpace = false);
 	static unsigned int createCubemapFromFile(const std::vector<std::string>& paths);
 	static unsigned int createTextureFromData(const TextureData & textureData);
 	static GLenum getTextureFormatFromData(const TextureData & textureData);
 	static GLenum getTextureFormatFromData(int componentCount);
-	static void SaveImage(const std::string& path, const glm::vec2& imageRes, ImageFormat imageFormat, char * data);
+	static void SaveImage(const std::string& path, const glm::ivec2& imageRes, ImageFormat imageFormat, char * data);
 	static unsigned int createTextureFromColour(const ColourData& colour);
 };
